Before:
```
benchmarking normalization of examples/Reducer.hs
time                 691.2 ms   (649.2 ms .. NaN s)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 676.9 ms   (661.8 ms .. 685.8 ms)
std dev              15.05 ms   (5.252 ms .. 20.46 ms)
variance introduced by outliers: 19% (moderately inflated)
```

After:
```
benchmarking normalization of examples/Reducer.hs
time                 542.3 ms   (449.8 ms .. 588.6 ms)
                     0.997 R²   (0.991 R² .. 1.000 R²)
mean                 517.8 ms   (496.6 ms .. 529.4 ms)
std dev              20.39 ms   (2.907 ms .. 26.71 ms)
variance introduced by outliers: 19% (moderately inflated)
```